+++ CHANGELOG.md
@@ -10,6 +10,7 @@ below this description) is likely unreleased.
 
 ## Added
 
+- Added `CmdContext.Stdin` since I already have `stderr` and `stdout`
 - Added `go.bbkane.com/warg/set` package with a basic set implementation
 - Added `warg.AllowForwardedArgs` to allow apps to forward args to the `Cmd`. Useful to create apps that wrap other CLI tools